backend/requirements.txt
# Core Python packages
setuptools>=80.0.0

# Django Core
Django==5.1
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Django Allauth (Social Authentication)
django-allauth==0.57.0

# Django Channels (WebSocket support)
channels==4.0.0
channels-redis==4.1.0

# Image Processing
Pillow==11.1.0

# PDF Generation
weasyprint==60.2

# M-Pesa Integration (optional - only if using M-Pesa payments)
# django-daraja==1.0.0

# Database
# For production, consider PostgreSQL for better concurrency and reliability

# Utilities
requests==2.32.3
PyJWT==2.10.1
